pyproject.toml
[tool.poetry]
name = "firebase-storage-utils"
version = "0.1.0"
description = "Python utility for Firebase Storage (Google Cloud Storage)."
authors = ["quwac <53551867+quwac@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.6.1"
google-cloud-storage = "^1.35.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
autopep8 = "^1.5.4"
black = "^20.8b1"
isort = "^5.6.4"
flake8 = "^3.8.4"
hacking = "^4.0.0"
flake8-annotations = "^2.4.1"
flake8-print = "^3.1.4"
flake8-variables-names = "^0.0.3"
flake8-simplify = "^0.11.0"
flake8-use-fstring = "^1.1"
fire = "^0.3.1"
json5 = "^0.9.5"
simple-chalk = "^0.1.0"
pytest-cov = "^2.10.1"
PyYAML = "^5.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
force_sort_within_sections = true
group_by_package = true

[tool.black]
line-length = 99

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v --cov=src --cov-branch --cov-report=html:lint_result/pytest"
